Add support for and update to Jackson 2.15.x
Jackson added the "JsonParser#getNumberValueDeferred()" method, including
a default implementation in "ParserBase". When using polymorphic types,
the default implementation returns an empty string for a double value.

This change introduces an implementation for "#getNumberValueDeferred"
in "JsonParserAdapter" to fix the problem.

Add a test for parsing and generating polymorphic types.

Fixes #236